requirements.txt
SQLAlchemy~=2.0.30
uvicorn~=0.31.1
fastapi~=0.115.2
requests~=2.32.3
opencv-python~=4.10.0.84
numpy~=2.1.2
pillow~=10.4.0
scikit-learn~=1.5.0
matplotlib~=3.9.0
jsonargparse~=4.31.0
anomalib~=1.1.0
openvino~=2024.2.0
docstring_parser~=0.16
rich~=13.7.1
lightning~=2.3.1
torch~=2.4.1+cu124
torchvision~=0.18.0+cu121
pandas~=2.2.2
tqdm~=4.66.4
tifffile~=2024.6.18
imgaug~=0.4.0
scipy~=1.13.0
omegaconf~=2.3.0
kornia~=0.7.3
scikit-image~=0.24.0
torchmetrics~=1.4.0.post0
einops~=0.8.0
FrEIA~=0.2
mpmath~=1.3.0
timm~=1.0.7
wandb~=0.17.3
PyYAML~=6.0.1
pytest~=8.2.0
groundingdino~=0.1.0
huggingface-hub~=0.23.4
typer~=0.12.3
transformers~=4.42.3
addict~=2.4.0
yapf~=0.40.2
supervision~=0.21.0
pycocotools~=2.0.8
setuptools~=75.2.0
seaborn~=0.13.2
loguru~=0.7.2
ipython~=8.28.0
traitlets~=5.14.3
SQLAlchemy-Utils~=0.41.2
CoilDataBase~=0.0.1
ultralytics~=8.3.12
colorlog~=6.8.2
XlsxWriter~=3.2.0
psutil~=6.0.0
GPUtil~=1.4.0
pymssql~=2.3.1
PyMySQL~=1.1.1